解决STreeView刷新单项问题。调整ITvDataSetObserver::OnBranchInvalidate方法，增加invalidChildren参数，默认为false。

+++ components/imgdecoder-gdip/imgdecoder-gdip.cpp
@@ -140,7 +140,7 @@ namespace SOUI
         // Get the number of frames in the first dimension.
         m_nFrameCount = max(1, bmpSrc->GetFrameCount(&pageGuid));
         
-        SIZE imSize={bmpSrc->GetWidth(),bmpSrc->GetHeight()};
+        SIZE imSize={(LONG)bmpSrc->GetWidth(),(LONG)bmpSrc->GetHeight()};
 
         if (m_nFrameCount>1)
         {
